% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_cmep}
\alias{parse_cmep}
\title{Parse CMEP Records}
\usage{
parse_cmep(recs, unnest = FALSE)
}
\arguments{
\item{recs}{character vector with CMEP formatted records}

\item{unnest}{boolean; should the data values be unnested? if FALSE (the
default) then they are returned in a list column with nested tables}
}
\value{
a [tibble][tibble::tibble-package] with the parsed records
}
\description{
Parse an individual CMEP records.
}
\examples{
f <- system.file("extdata", "cmep.dat", package = "cmep", mustWork = TRUE)
cmep_f <- read_cmep(f, parse = FALSE)
parse_cmep(cmep_f)
}
