% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_cmep}
\alias{read_cmep}
\alias{read_cmep_chunked}
\title{Read CMEP File}
\usage{
read_cmep(..., parse = TRUE, unnest = FALSE)

read_cmep_chunked(..., parse = TRUE, unnest = FALSE)
}
\arguments{
\item{...}{passed to \code{\link[readr]{read_lines}} or
\code{\link[readr]{read_lines_chunked}}}

\item{parse}{boolean; should the records be parsed?}

\item{unnest}{boolean; should the data values be unnested? if FALSE (the
default) then they are returned in a list column with nested tables}
}
\description{
Read and parse CMEP file. These functions are mostly just simple wrappers
around the more generic \code{readr} functions. The main enhancement is that
they will parse an irregularly sized CMEP file and format the data as a
regular table (data frame). \code{read_cmep_chunked} can be used to process
files that may be too large to fit in memory.
}
\examples{
f <- system.file("extdata", "cmep.dat", package = "cmep", mustWork = TRUE)
read_cmep(f)
f <- system.file("extdata", "cmep.dat", package = "cmep", mustWork = TRUE)
cb_f <- function(x, pos) print(pos)
read_cmep_chunked(f, callback = cb_f, chunk_size = 2)
}
