% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_cmep}
\alias{read_cmep}
\title{Read CMEP File}
\usage{
read_cmep(..., parse = TRUE)
}
\arguments{
\item{...}{passed to \code{\link[readr]{read_lines}}}

\item{parse}{boolean; should the records be parsed?}
}
\description{
Read and parse CMEP file. Currently, the functions are just simple wrappers
around the more generic \code{readr} functions.
}
\examples{
f <- system.file("extdata", "cmep.dat", package = "cmep", mustWork = TRUE)
read_cmep(f)
}
