% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_cmep}
\alias{read_cmep}
\title{Read CMEP File}
\usage{
read_cmep(file, ...)
}
\arguments{
\item{file}{path to file}

\item{...}{passed to \code{\link[readr]{read_lines}}}
}
\description{
Read and parse CMEP file
}
\examples{
f <- system.file("extdata", "cmep.dat", package = "cmep", mustWork = TRUE)
read_cmep(f)
}
